New  update add entity classes and modifies

+++ src/main/java/com/accenture/lkm/dao/EmployeeDAO.java
@@ -6,15 +6,15 @@ import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
 import org.springframework.transaction.annotation.Transactional;
 
-import com.accenture.lkm.entity.EmployeeEntity;
+
 import com.accenture.lkm.entity.Person;
 
-public interface EmployeeDAO  extends JpaRepository<Person, Integer>{
+public interface EmployeeDAO  extends JpaRepository<Person, Long>{
 	
 	@Transactional
-	@Modifying
+	 @Modifying
 	@Query("delete from Person p  where p.U_id=:id")
-	int delete(@Param("id") long id);
+	int delete(@Param("id") int id);
 	
 	@Query("select p from Person p  where p.U_id=:id")
 	Person findById(@Param("id") long id);
